+++ packages/Chem/src/tests/menu-tests-script-based.ts
@@ -34,7 +34,7 @@ main_component_non_st,CCC1=C(C)C=CC(O)=N1`);
   });
 
   test('curate.smiles', async () => {
-    await curate(DG.Test.isInBenchmark ? grok.data.demo.molecules(500) : smiles, 'smiles');
+    await curate(DG.Test.isInBenchmark ? grok.data.demo.molecules(1000) : smiles, 'smiles');
   }, {timeout: 60000, benchmark: true});
 
   test('curate.molV2000', async () => {