+++ packages/Chem/src/tests/rendering-tests.ts
@@ -48,7 +48,7 @@ category('rendering', () => {
   });
 
   test('rdkit grid cell renderer scroll', async () => {
-    const df = DG.Test.isInBenchmark ? await grok.data.files.openTable('Demo:Files/chem/smiles_1M.zip') :
+    const df = DG.Test.isInBenchmark ? await grok.data.files.openTable('System:DemoFiles/chem/smiles_1M.zip') :
       await readDataframe('tests/sar-small_test.csv');
     const scrollCycles = DG.Test.isInBenchmark ? 1000 : 10;
     const scrollDelta = DG.Test.isInBenchmark ? 150000 : 10000;